Finished scrollable icon panel
git-svn-id: https://appetizer.googlecode.com/svn@376 ff61cf6c-2836-11df-8847-f3d5e5ef3dde

+++ branches/QtVersion/gui/ScrollBar.cpp
@@ -77,7 +77,7 @@ int ScrollBar::defaultWidth() const {
 
 int ScrollBar::knobHeight() {
   if (!scrollable()) return 0;
-  int output = (height() / contentHeight_) * height();
+  int output = (int)(((float)height() / (float)contentHeight_) * (float)height());
   if (output < 20) output = 20;
 	return output;
 }
